patient field modification done

+++ hms/src/main/java/com/hms/clinic/suyesha/dto/PatientDto.java
@@ -23,7 +23,7 @@ public class PatientDto{
 	
 	private String address;
 	
-	private Integer mobile;
+	private Long mobile;
 	
 	private String cghsPrivate;
